- Multiple small code cleanups, mostly unsigned/signed issues, with a few variable size redefinitions.
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@3990 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/common/grfio.c
@@ -818,7 +818,7 @@ static char * decode_filename(unsigned char *buf,int len)
 static int grfio_entryread(char *gfname,int gentry)
 {
 	FILE *fp;
-	int grf_size,list_size;
+	long grf_size,list_size;
 	unsigned char grf_header[0x2e];
 	int lop,entry,entrys,ofs,grf_version;
 	char *fname;
