+++ src/map/npc.c
@@ -1029,7 +1029,7 @@ int npc_buylist(struct map_session_data *sd,int n,unsigned short *item_list)
 
 	//Logs (S)hopping Zeny [Lupus]
 	if(log_config.zeny > 0 )
-		log_zeny(sd, "S", sd, -z);
+		log_zeny(sd, "S", sd, -(int)z);
 	//Logs
 
 	pc_payzeny(sd,(int)z);
@@ -1096,7 +1096,7 @@ int npc_selllist(struct map_session_data *sd,int n,unsigned short *item_list)
 
 	//Logs (S)hopping Zeny [Lupus]
 	if(log_config.zeny > 0 )
-		log_zeny(sd, "S", sd, z);
+		log_zeny(sd, "S", sd, (int)z);
 	//Logs
 
 	pc_getzeny(sd,(int)z);
